auth rework
