+++ client/src/app/admin/login/setCookie.ts
@@ -1,0 +1,14 @@
+"use server";
+
+import { cookies } from "next/headers";
+
+export default async function setCookie(token: string) {
+  cookies().set({
+    name: "token",
+    value: token,
+    httpOnly: true,
+    secure: true,
+    path: "/",
+  });
+  console.log(token);
+}
